Fixed a bug when writing the cache.
After a previous edit to the structure where BSFRC doesn't store an array of sections, only a dictionary of them in _secionsByName, the code that writes the cache needed to be updated.

Currently the cache gets written after every change notification (and after the initial fetch). This might be too frequent, but exactly when to write the cache is a non-trivial problem.

Signed-off-by: Daniel Thorpe <dan@blindingskies.com>